Transfer of "Personal-Website-Project" Repository
I mistakenly created a completely new repository for the project and
uploaded it on time, before the due date on GitHub. After attending
class today, I realized that it was not properly uploaded under my
jacquelinejwilson.github.io so I have uploaded the exact same code to
this repository so my personal website would be in the correct
location. Again, I will stress that this code is exactly the same as
the other repository that was submitted on time before the due date, it
has not been altered in any way.